+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -103,5 +103,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-12T09:11:18.701Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-12T10:14:30.867Z")
     r"""The date and time the resource was last updated"""