+++ packages/orq-rc/RELEASES.md
@@ -789,3 +789,13 @@ Based on:
 - [python v3.2.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.41] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.41 - packages/orq-rc
+
+## 2025-03-12 10:15:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.517.0 (2.548.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.42] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.42 - packages/orq-rc